Refactor NavbarContent
