+++ src/components/navbar/RegularNavbar.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import {connect} from 'react-redux';
-
 import {stateTypes, dispatchers} from 'utilities';
-
-import NavbarContent from './components/NavbarContent';
+import NavbarContent from './navbar-content/navbar-content';
 
 interface RegularNavbarProps {
     logOut(): void;
