+++ src/components/navbar/navbar-content/visual-navbar-button.tsx
@@ -1,17 +1,19 @@
 import React from 'react';
 
-interface NavbarLinkComponentProps {
+interface VisualNavbarButtonProps {
     text: string;
     active?: boolean;
     icon: React.ReactNode;
     onClick?(): void;
 }
-function NavbarButton(props: NavbarLinkComponentProps) {
+function VisualNavbarButton(props: VisualNavbarButtonProps) {
     return (
         <div
-            onClick={props.onClick ? props.onClick : () => {}}
+            onClick={props.onClick}
             className={
-                'relative w-58 h-12 mx-3 my-1 px-2 py-1 text-lg font-weight-600 flex flex-row items-center justify-start rounded cursor-pointer ' +
+                'relative w-58 h-12 mx-3 my-1 px-2 py-1 text-lg ' +
+                'font-weight-600 flex flex-row items-center ' +
+                'justify-start rounded cursor-pointer ' +
                 (props.active
                     ? 'text-white bg-gray-800'
                     : 'text-gray-400 hover:text-white hover:bg-gray-800')
@@ -23,4 +25,4 @@ function NavbarButton(props: NavbarLinkComponentProps) {
     );
 }
 
-export default NavbarButton;
+export default VisualNavbarButton;
